--- a/src/GraphQL/Execution/SubscriptionExecutionStrategy.cs
+++ b/src/GraphQL/Execution/SubscriptionExecutionStrategy.cs
@@ -103,7 +103,7 @@ namespace GraphQL.Execution
                 return subscription
                     .Select(value =>
                     {
-                        var executionNode = BuildExecutionNode(node.Parent, node.GraphType, node.Field, node.FieldDefinition, node.IndexInParentNode);
+                        var executionNode = BuildExecutionNode(node.Context, node.Parent, node.GraphType, node.Field, node.FieldDefinition, node.IndexInParentNode);
                         executionNode.Source = value;
                         return executionNode;
                     })