This PR makes ExecutionNode implement IResolveFieldContext, so an instance of ReadonlyResolveFieldContext is not required, saving an object allocation per resolver.  Doesn't really enhance performance, although the number of object allocations decrease notably as can be seen in the GC stats below.  Memory doesn't change much due to the previous enhancement of the ReadonlyResolveFieldContext over the prior ResolveFieldContext class.

Just a thought - ok to merge or close.

```
// * Summary *

BenchmarkDotNet=v0.12.0, OS=Windows 10.0.18362
Intel Core i7-7740X CPU 4.30GHz (Kaby Lake), 1 CPU, 8 logical and 4 physical cores
.NET Core SDK=3.1.101
  [Host]     : .NET Core 3.1.1 (CoreCLR 4.700.19.60701, CoreFX 4.700.19.60801), X64 RyuJIT
  DefaultJob : .NET Core 3.1.1 (CoreCLR 4.700.19.60701, CoreFX 4.700.19.60801), X64 RyuJIT


*** BEFORE ***

|        Method |        Mean |     Error |    StdDev |    Gen 0 |   Gen 1 | Gen 2 | Allocated |
|-------------- |------------:|----------:|----------:|---------:|--------:|------:|----------:|
| Introspection | 1,503.10 us | 19.721 us | 17.483 us | 189.4531 | 60.5469 |     - | 835.56 KB |
|          Hero |    34.32 us |  0.639 us |  0.566 us |   6.1035 |  0.0610 |     - |   25.1 KB |

*** AFTER ***

|        Method |        Mean |     Error |    StdDev |    Gen 0 |   Gen 1 | Gen 2 | Allocated |
|-------------- |------------:|----------:|----------:|---------:|--------:|------:|----------:|
| Introspection | 1,496.35 us | 19.890 us | 18.605 us | 171.8750 | 48.8281 |     - | 807.63 KB |
|          Hero |    34.44 us |  0.597 us |  0.558 us |   6.1035 |       - |     - |  25.05 KB |
```