@sungam3r 2020-02-16
I am familiar with such optimization methods and I have an ambiguous attitude towards them. On the one hand, the number of entities that work in close interconnection is reduced. It really saves memory. On the other hand, combining entities into one can affect code clarity and problems during project evolution.

 I looked at the code briefly and will still watch it again. So far my opinion is 50/50.