@Shane32 2020-03-10
Nope.  I don't know why, but it's always been that way.  My PC always tests in X64 and your baseline comparison tests are always X86.