@Shane32 2020-03-10
The last time I did some testing, I found that someone had changed the default tests to run speed comparisons with the different serializer packages, so I did have to reconfigure it to run the Introspection and Hero tests.