@Shane32 2020-08-14
@sungam3r Let me know your general thoughts on this PR.  I'm happy to close it or update it, either way.  I doubt this will provide any meaningful benefit for anyone, even if it is a measurable speed increase.