@sungam3r 2020-08-14
The difference in performance is not that big. I have a suspicion that such a merge might cause then problems for us with DI evolution after green light to #1730. There probably will be new `IResolveFieldContext` implementations. As I said earlier
> On the other hand, combining entities into one can affect code clarity and problems during project evolution.

Closing.